Reading as bytes

+++ src/map_builder.rs
@@ -1,5 +1,5 @@
-pub fn map_builder() -> Vec<u16> {
-  let mut v: Vec<u16> = Vec::new();
+pub fn map_builder() -> Vec<u8> {
+  let mut v: Vec<u8> = Vec::new();
   v.reserve(81);
   v
 }
